chore: update ruby-lsp runtime dependency
Removed ruby-lsp runtime dependency to make it possible to work with any
compatible version without locking the version in the gemspec file. With
the new RubyLsp depend_on_ruby_lsp! method it is possible to be forward
compatible with non breaking change versions and be disabled if the
chore lsp updates its API.